Moved function _define_problem_with_groups to util, since it can be used by both sample and analyze packages
